fix: cast float division result to uint for AwaitMillis
UITimings constants are uint, so division result must be cast back to uint
before passing to AwaitMillis method signature.

+++ scene_tests/ScoundrelSceneTests.cs
@@ -535,7 +535,7 @@ public class ScoundrelSceneTests
         var weapon = FindRoomCard(scene, s => s == "diamonds");
         AssertThat(weapon).IsNotNull();
         ClickCard(scene, weapon!);
-        await _runner!.AwaitMillis(UITimings.DragAnimationMs / 1.5f); // wait shorter since weapon animates briefly
+        await _runner!.AwaitMillis((uint)(UITimings.DragAnimationMs / 1.5f)); // wait shorter since weapon animates briefly
 
         AssertThat(scene.GetNode<Label>("UI/LeftPanel/WeaponGroup/WeaponLabel").Text).IsNotEqual("Weapon: none");
